requirements.txt
PyYAML>=6.0
requests>=2.31.0
playwright>=1.45.0
Pillow>=10.0.0
numpy>=1.26.0
google-api-python-client>=2.130.0
google-auth>=2.23.0,<2.24.0
google-auth-oauthlib>=1.2.0
google-cloud-storage>=2.16.0
tzdata>=2024.1

PySide6>=6.8.0
reportlab>=4.0.0
